updated with-context-api example to utilize the App Router.
