+++ examples/with-context-api/app/layout.tsx
@@ -1,0 +1,20 @@
+import { CounterProvider } from "../app/_components/Counter";
+
+export const metadata = {
+  title: 'Next.js',
+  description: 'Generated by Next.js',
+}
+
+export default function RootLayout({
+  children,
+}: {
+  children: React.ReactNode
+}) {
+  return (
+    <CounterProvider>
+      <html lang="en">
+        <body>{children}</body>
+      </html>
+    </CounterProvider>
+  )
+}
