--- examples/with-context-api/pages/_app.tsx
@@ -1,10 +1,0 @@
-import type { AppProps } from "next/app";
-import { CounterProvider } from "../components/Counter";
-
-export default function MyApp({ Component, pageProps }: AppProps) {
-  return (
-    <CounterProvider>
-      <Component {...pageProps} />
-    </CounterProvider>
-  );
-}
